requirements.txt
prettytable==0.7.2
moviepy==1.0.3
numpy==1.18.2
matplotlib==3.3.0
ffmpeg-python==0.2.0